src/main/webapp/app/entities/statement/route/statement-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IStatement, Statement } from '../statement.model';
import { StatementService } from '../service/statement.service';

import { StatementRoutingResolveService } from './statement-routing-resolve.service';

describe('Statement routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: StatementRoutingResolveService;
  let service: StatementService;
  let resultStatement: IStatement | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(StatementRoutingResolveService);
    service = TestBed.inject(StatementService);
    resultStatement = undefined;
  });

  describe('resolve', () => {
    it('should return IStatement returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 'ABC' };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultStatement = result;
      });

      // THEN
      expect(service.find).toBeCalledWith('ABC');
      expect(resultStatement).toEqual({ id: 'ABC' });
    });

    it('should return new IStatement if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultStatement = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultStatement).toEqual(new Statement());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Statement })));
      mockActivatedRouteSnapshot.params = { id: 'ABC' };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultStatement = result;
      });

      // THEN
      expect(service.find).toBeCalledWith('ABC');
      expect(resultStatement).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
